Fix additional warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@1191 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-crt/profile/gmon.c
@@ -80,7 +80,7 @@ monstartup(lowpc, highpc)
 	size_t lowpc;
 	size_t highpc;
 {
-	register int o;
+	register size_t o;
 	char *cp;
 	struct gmonparam *p = &_gmonparam;
